requirements.txt
streamlit==1.40.1 
pandas==1.5.3
numpy==1.24.4
matplotlib==3.4.3
seaborn==0.11.2
Pillow==10.0.1
requests==2.32.3
tqdm==4.67.1
xhtml2pdf==0.2.17
XlsxWriter==3.2.1
pyarrow